requirements.txt
pyserial
crcmod
influxdb

pytest
